Fix config flow issues
The `application_credentials` needs to come before the custom steps for
the integration. When credentials need to be added, it seems as though
it basically kills the current flow, allows the creation of credentials,
and starts a new flow. So none of the existing state gets saved.

These issues were reported in #35.