Fix location restore state
The restore method checks that the entity is not yet available at the
time of restore (to avoid clobbering any changed state that happened at
startup). But the location's entity description has a value cast that
ensures the extracted value is always a dictionary. So testing against
the final (casted) value always results in the entity being `available`.
Therefore, availability should be based on whether the `raw_value` is
non-null.

Resolves: #36