Improve availability check
This is a followup to #37 to improve the availability check. In fact,
for entities that share coordinator data, i.e. tire pressure, it will be
better to ensure that both the raw_value and value are not null. This
will ensure that not just the first entity's data data gets restored.